Moved transcode/upload to lib/objectstore
