+++ lib/objectstore/go.mod
@@ -1,0 +1,3 @@
+module gitlab.liu.se/adaab301/tddd27_2022_project/lib/objectstore
+
+go 1.18
