+++ transcode/transcoder/h264/h264.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 
 	ffmpeg_go "github.com/u2takey/ffmpeg-go"
+	"gitlab.liu.se/adaab301/tddd27_2022_project/lib/objectstore"
 	"gitlab.liu.se/adaab301/tddd27_2022_project/transcode/fileutil"
-	"gitlab.liu.se/adaab301/tddd27_2022_project/transcode/upload"
 )
 
 func TranscodeToh264(fileName string, originalFileName string, dir string, uid string) error {
@@ -17,6 +17,6 @@ func TranscodeToh264(fileName string, originalFileName string, dir string, uid s
 		Run()
 
 	dirName := fileutil.RemoveFileNameFromDirectory(dir)
-	upload.FiletoDB(dirName+"/"+"video"+".mp4", outputFileName, originalFileName, uid)
+	objectstore.FiletoDB(dirName+"/"+"video"+".mp4", outputFileName, originalFileName, uid)
 	return nil
 }
